requirements.txt
langchain-ollama
tavily-python
typing-extensions
langchain-core
langgraph
Flask